Build with Visual Studio 2022 (v143), zero errors and warnings
Migrate the solution from VS2013 (v120) to VS2022 (v143) Release x64.

Modernize removed STL (/Zc:forScope-, hash_map, bind1st/2, binary_function, export methods), fix /GL+LTCG inline-emission link gaps (FS_impl_anchor.cpp, de-inlined single-TU IC defs, missing *_impl.h includes), complete-enum includes for luabind C2027, drop bogus XRCORE_API from _quaternion<T>. Verified: full build 0 errors/0 warnings, game boots to main menu. See CHANGELOG.md.

+++ sdk/include/vorbis/lib/vorbisfile.c
@@ -68,7 +68,7 @@ static long _get_data(OggVorbis_File *vf){
   if(!(vf->callbacks.read_func))return(-1);
   if(vf->datasource){
     char *buffer=ogg_sync_buffer(&vf->oy,CHUNKSIZE);
-    long bytes=(vf->callbacks.read_func)(buffer,1,CHUNKSIZE,vf->datasource);
+    long bytes=(long)(vf->callbacks.read_func)(buffer,1,CHUNKSIZE,vf->datasource);
     if(bytes>0)ogg_sync_wrote(&vf->oy,bytes);
     if(bytes==0 && errno)return(-1);
     return(bytes);
